requirements.txt
kafka-python
SQLAlchemy
mysqlclient
grpcalchemy
grpcio-tools
grpc_tools